+++ docs/history.md
@@ -48,6 +48,7 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 - Allow build data to add additional project dependencies for `wheel` and `sdist` build targets
 - Add `force_include_editable` build data for the `wheel` build target
 - Add `build_hooks` build data
+- Add support for Mercurial's `.hgignore` files when using glob syntax
 - Update project metadata to reflect the adoption by PyPA
 
 ***Fixed:***